1. Fix the build error caused by the modification for compiler 6. 2. Revise startup_NUC131.s and retarget.c files to fix build issue. 3. Modify settings of Keil project to speed up compilation. 4. Revise copyright format for gen. chm file. 5. Remove useless files of IAR Project.

+++ SampleCode/StdDriver/ADC_PwmTrigger/main.c
@@ -5,8 +5,9 @@
  * $Date: 14/06/30 4:51p $
  * @brief    Demonstrate how to trigger ADC by PWM.
  * @note
- * Copyright (C) 2014 Nuvoton Technology Corp. All rights reserved.
+ * @copyright SPDX-License-Identifier: Apache-2.0
  *
+ * @copyright Copyright (C) 2014 Nuvoton Technology Corp. All rights reserved.
  ******************************************************************************/
 #include <stdio.h>
 #include "NUC131.h"
@@ -185,7 +186,7 @@ void ADC_PWMTrigTest_SingleOpMode()
     /* Clear the ADC interrupt flag */
     ADC_CLR_INT_FLAG(ADC, ADC_ADF_INT);
 
-    printf("Channel 2: 0x%X\n", ADC_GET_CONVERSION_DATA(ADC, 2));
+    printf("Channel 2: 0x%X\n", (unsigned int)ADC_GET_CONVERSION_DATA(ADC, 2));
 
     /* Disable ADC */
     ADC_POWER_DOWN(ADC);
